Atualização Modulo01

+++ Modulo01/Pontos2D-C/Ponto2D.c
@@ -4,6 +4,6 @@
 float distancia(Ponto2D p1, Ponto2D p2) {
     float dx = p1.x - p2.x;
     float dy = p1.y - p2.y;
-    return sqrt (dx * dx + dy * dy);
+    return ((float)sqrt (dx * dx + dy * dy));
 }
